% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_frac_mq}
\alias{copy_frac_mq}
\title{Copy \code{frac_smry.xlsx}}
\usage{
copy_frac_mq(dat_dir, from = "frac_smry.xlsx", to = "frac_smry.xlsx")
}
\description{
\code{copy_frac} copies the \code{frac_smry.xlsx} to a target directory.
}
