% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCPTAC.R
\name{copy_mq_msms}
\alias{copy_mq_msms}
\title{Copy MaxQuant \code{.txt} files}
\usage{
copy_mq_msms(dat_dir, filelist)
}
\description{
\code{copy_mq_msms} copies the MaxQuant \code{msms.txt} files to a target
directory.
}
